// src/app/main/models/notification.ts
export interface INotification {
    title: string;
    message: string;

    aggId: string;
    application: string;
    // authStatus: null
    // checkerDate: null
    // checkerId: null
    createdTime: number;
    from: string;
    id: number;
    link: any;
    makerDate: number;
    makerId: any;
    object: any;
    recordStatus: string;
    to: string;
    unread: number;
    updatedTime: number;
    user: string;
    isReaded: number;
    time: string;
}